refactor: bring three oversized files under the 800-LOC policy
The file-size gate flagged three files that grew past 800 LOC while
the verify gates were skipped. Restructure instead of adding
exceptions:

- uffs-bench env.rs (848): extract the inline test module to
  env/tests.rs, the established sibling-module pattern (539 + 311).
- uffs-mft commands/windows/info.rs (986): split the independent
  drives command (cmd_drives + DriveInfo + DriveRecord + printers +
  get_volume_label) into drives.rs (705 + 315); each file keeps only
  the lint expectations its half fulfils.
- uffs-bench run/mod.rs (818): extract the pure stage identity +
  selection helpers (stage_selected, stage_step_id, stage_banner,
  stage0_outputs) to run/stage_meta.rs (780 + 56).

No behavior change. Validated with host clippy/nextest (96 passed)
and cargo xwin clippy for the Windows-gated split.
